sync GitLab with Azure

+++ src/WorkforeManagement.BLL/Services/UserService.cs
@@ -2,11 +2,10 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Security.Claims;
-using System.Text;
 using System.Threading.Tasks;
+using WorkforceManagement.BLL.IHelpers;
 using WorkforceManagement.BLL.IServices;
 using WorkforceManagement.DAL.Entities;
-using WorkforceManagement.DAL.IRepositories;
 using WorkforceManagement.DAL.Repositories;
 
 namespace WorkforceManagement.BLL.Services
@@ -15,15 +14,18 @@ namespace WorkforceManagement.BLL.Services
     {
         private readonly IUserManager _userManager;
         private readonly ITimeOffRequestRepository _timeOffRequestRepository;
+        private readonly IUserServiceHelper _helper;
 
-        public UserService(IUserManager userManager, ITimeOffRequestRepository timeOffRequestRepository)
+        public UserService(IUserManager userManager, ITimeOffRequestRepository timeOffRequestRepository, IUserServiceHelper helper)
         {
             _userManager = userManager;
             _timeOffRequestRepository = timeOffRequestRepository;
+            _helper = helper;
         }
-        public async Task<User> GetCurrentUser(ClaimsPrincipal principal)
+
+        public async Task<User> GetUserByIdAsync(string userId)
         {
-            return await _userManager.GetCurrentUser(principal);
+            return await _userManager.FindByIdAsync(userId);
         }
         public async Task<User> GetUserByNameAsync(string userName)
         {
@@ -37,59 +39,47 @@ namespace WorkforceManagement.BLL.Services
         {
             return await _userManager.GetAllAsync();
         }
-        public async Task<bool> CreateUserAsync(string userName, string eMail, string passWord, string firstName, string lastName, Role role)
+        public async Task<User> GetCurrentUserAsync(ClaimsPrincipal principal)
         {
-            if (await _userManager.FindByNameAsync(userName) != null) return false;
-            if (await _userManager.FindByEmailAsync(eMail) != null) return false;
-
-            User newUser = new()
-            {
-                UserName = userName,
-                Email = eMail,
-                FirstName = firstName,
-                LastName = lastName,
-                CreatedAt = DateTime.Now
-            };
-
-            await _userManager.CreateUserAsync(newUser, passWord);
-            await AddUserToRoleAsync(newUser, role);
-            return true;
+            return await _userManager.GetCurrentUser(principal);
         }
-        public async Task<bool> EditUserAsync(string userId, string userName, string eMail, string currentPassword, string newPassword, string firstName, string lastName)
+        public async Task<User> CreateUserAsync(User newUser, string passWord, string role)
         {
-            if (await _userManager.FindByNameAsync(userName) != null) return false;
-            if (await _userManager.FindByEmailAsync(eMail) != null) return false;
+            await _helper.CheckDuplicateEmailAndUsernameAsync(newUser);
 
-            User userToBeEdited = await _userManager.FindByIdAsync(userId);
+            newUser.CreatedAt = DateTime.Now;
+            role = _helper.FormatRole(role);
 
-            userToBeEdited.UserName = userName;
-            userToBeEdited.FirstName = firstName;
-            userToBeEdited.LastName = lastName;
-            userToBeEdited.Email = eMail;
+            newUser = await _userManager.CreateUserAsync(newUser, passWord);
+            await _userManager.AddToRoleAsync(newUser, role);
+
+            return newUser;
+        }
+        public async Task EditUserAsync(User modelWithUpdates, string currentPassword, string newPassword)
+        {
+            await _helper.CheckDuplicateEmailAndUsernameAsync(modelWithUpdates);
+
+            var userToBeEdited = await _userManager.FindByIdAsync(modelWithUpdates.Id);
+
+            userToBeEdited.UserName = modelWithUpdates.UserName;
+            userToBeEdited.FirstName = modelWithUpdates.FirstName;
+            userToBeEdited.LastName = modelWithUpdates.LastName;
+            userToBeEdited.Email = modelWithUpdates.Email;
 
             await _userManager.UpdateUserAsync(userToBeEdited, currentPassword, newPassword);
-            return true;
         }
         public async Task DeleteUserAsync(string userId)
         {
             var userToBeDeleted = await _userManager.FindByIdAsync(userId);
-            var userTimeOffs = await _timeOffRequestRepository.GetAllTimeOffsByUser(userToBeDeleted);
-            await _timeOffRequestRepository.DeleteCollectionAsync(userTimeOffs);
-            // TODO: Delete related information such as TimeOffRequests. Currently not working for users that have created TimeOffRequests
-            // _timeOffRequestRepository.DeleteCollection(userToBeDeleted.TimeOffRequests)
-            await _userManager.DeleteUserAsync(userToBeDeleted);
-        }   
+            if (_userManager.GetUserRolesAsync(userToBeDeleted).Result.Contains("Admin"))
+                throw new ArgumentException("You can not delete Admin users! ");
 
-        public async Task AddUserToRoleAsync(User user, Role role)
-        {
-            if (role == Role.Admin)
-            {
-                await _userManager.AddToRoleAsync(user, "Admin");
-            }
-            else
-            {
-                await _userManager.AddToRoleAsync(user, "Regular");
-            }
-        }
+            var userTimeOffs = await _timeOffRequestRepository.GetAllTimeOffsByUser(userToBeDeleted);
+            await _timeOffRequestRepository.DeleteCollectionAsync(userTimeOffs.ToList());
+
+            await _helper.CheckIfUserIsTeamLeaderAsync(userToBeDeleted);
+
+            await _userManager.DeleteUserAsync(userToBeDeleted);
+        }        
     }
 }
